Remap mode-line face instead of inverting it
There are times, when running an Emacs server, that a new frame incorrectly has
an inverted mode-line face.  I think it has to do with a race condition when
inverting the face.

To correct this issue, I've added a face `mode-line-bell`, which inherits from
the `mode-line` face and inverts the video. `mode-line-flash` now adds and
removes that face from `face-remapping-alist`, and the `mode-line-bell-mode`
also adds a function to `after-focus-change-function` to recover the correct
mode-line face.

The main issue with this correction is that the inverse-video property on
`mode-line-bell` behaves slightly differently than the previous functionality of
this mode.  Suggestions on how to solve this are welcome -- as far as I can
guess, the best way to do that would be to hook the defface call into changing
themes, which is clunky.