Initialize properly and hook into `after-make-frame-functions'
I'm still not sure what's causing the /original/ issue leading to PR #2, but
this should fix all the bugs that have been introduced in my fork so far.

To wit, the `mode-line-orig' face should only be created for a new frame and
after changing themes.  After the focus is changed, any pending switch should
unswitch.